Implement the basic data output for multiPopMcmc and singlePopMcmc
If applied, this commit will:

 - Begin fleshing out SqlBackingStore
 - Add sqlite3.[ch]
 - Add basic opening/closing of a database
 - Add first (dummy, hard-coded) sqlite invocation to singlePop
 - Add run and iteration tables
 - Extend the breadth of the SQL back-end
 - Add Settings entry for back end
 - Add backend to reference conf file
 - Add SQL insertion code for single/multiPop
 - Add sql script directory
 - Add SQL output for starParams

   This brings single and multiPopMcmc fully online, though they are currently only accessible via a SQLite CLI/GUI

+++ scatterCluster/scatterCluster.cpp
@@ -178,7 +178,7 @@ int main (int argc, char *argv[])
 
         if (settings.verbose)
         {
-            for (int f = 0; f < filters.size(); ++f)
+            for (size_t f = 0; f < filters.size(); ++f)
                 cout << f << ' ' << filters.at(f) << endl;
         }
